Adds EventDetails class w/ one-to-one relationship to Event
