+++ src/main/java/org/launchcode/codingevents/models/EventDetails.java
@@ -1,0 +1,44 @@
+package org.launchcode.codingevents.models;
+
+import javax.persistence.Entity;
+import javax.persistence.OneToOne;
+import javax.validation.constraints.Email;
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.Size;
+
+/**
+ * Created by Chris Bay
+ */
+@Entity
+public class EventDetails extends AbstractEntity {
+
+    @Size(max = 500, message = "Description too long!")
+    private String description;
+
+    @NotBlank(message = "Email is required")
+    @Email(message = "Invalid email. Try again.")
+    private String contactEmail;
+
+    public EventDetails(@Size(max = 500, message = "Description too long!") String description, @NotBlank(message = "Email is required") @Email(message = "Invalid email. Try again.") String contactEmail) {
+        this.description = description;
+        this.contactEmail = contactEmail;
+    }
+
+    public EventDetails() {}
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public String getContactEmail() {
+        return contactEmail;
+    }
+
+    public void setContactEmail(String contactEmail) {
+        this.contactEmail = contactEmail;
+    }
+}
