+++ src/main/resources/templates/events/index.html
@@ -18,8 +18,8 @@
     <tr th:each="event : ${events}">
         <td th:text="${event.id}"></td>
         <td th:text="${event.name}"></td>
-        <td th:text="${event.description}"></td>
-        <td th:text="${event.contactEmail}"></td>
+        <td th:text="${event.eventDetails.description}"></td>
+        <td th:text="${event.eventDetails.contactEmail}"></td>
         <td th:text="${event.eventCategory.name}"></td>
     </tr>
 </table>
